started on creating frontend and backend modules
