+++ backend/build.gradle
@@ -1,0 +1,39 @@
+
+plugins {
+    id 'java'
+    id 'org.springframework.boot' version '2.2.2.RELEASE'
+    id 'io.spring.dependency-management' version '1.0.8.RELEASE'
+}
+
+group = 'net.dancier'
+version = '0.0.1-SNAPSHOT'
+
+description = "Dancer"
+
+sourceCompatibility = 1.10
+targetCompatibility = 1.10
+tasks.withType(JavaCompile) {
+	options.encoding = 'UTF-8'
+}
+
+repositories {
+     mavenCentral()
+}
+
+tasks.named('processResources').configure {
+    dependsOn ':frontend:copyFrontendResources'
+}
+
+dependencies {
+    implementation 'org.projectlombok:lombok:1.18.12'
+    annotationProcessor 'org.projectlombok:lombok:1.18.12'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa:2.2.2.RELEASE'
+    implementation 'org.springframework.boot:spring-boot-starter-data-rest:2.2.2.RELEASE'
+    implementation  'org.springframework.boot:spring-boot-starter-thymeleaf:2.2.2.RELEASE'
+    implementation 'org.springframework.boot:spring-boot-starter-websocket:2.2.2.RELEASE'
+    implementation 'org.springframework.boot:spring-boot-starter-security:2.2.2.RELEASE'
+    implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity5:3.0.4.RELEASE'
+    runtime 'org.springframework.boot:spring-boot-devtools:2.2.2.RELEASE'
+    runtime 'com.h2database:h2:1.4.200'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test:2.2.2.RELEASE'
+}
