+++ settings.gradle
@@ -1,1 +1,4 @@
 rootProject.name = 'dancer'
+
+include 'frontend'
+include 'backend'
