--- src/main/java/de/frubumi/dance/dancer/DancerUser.java
@@ -1,97 +1,0 @@
-package de.frubumi.dance.dancer;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import java.util.Arrays;
-import java.util.Objects;
-
-@Entity
-public class DancerUser {
-
-    public static final PasswordEncoder PASSWORD_ENCODER = new BCryptPasswordEncoder();
-
-    private @Id
-    @GeneratedValue
-    Long id;
-
-    private String name;
-
-    private @JsonIgnore
-    String password;
-
-    private String[] roles;
-
-    public void setPassword(String password) {
-        this.password = PASSWORD_ENCODER.encode(password);
-    }
-
-    protected DancerUser() {}
-
-    public DancerUser(String name, String password, String... roles) {
-
-        this.name = name;
-        this.setPassword(password);
-        this.roles = roles;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        DancerUser DancerUser = (DancerUser) o;
-        return Objects.equals(id, DancerUser.id) &&
-                Objects.equals(name, DancerUser.name) &&
-                Objects.equals(password, DancerUser.password) &&
-                Arrays.equals(roles, DancerUser.roles);
-    }
-
-    @Override
-    public int hashCode() {
-
-        int result = Objects.hash(id, name, password);
-        result = 31 * result + Arrays.hashCode(roles);
-        return result;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public String[] getRoles() {
-        return roles;
-    }
-
-    public void setRoles(String[] roles) {
-        this.roles = roles;
-    }
-
-    @Override
-    public String toString() {
-        return "DancerUser{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", roles=" + Arrays.toString(roles) +
-                '}';
-    }
-}
